Extract class refactoring implemented
