+++ bytetcc-core/src/main/java/org/bytesoft/bytetcc/CompensableCoordinator.java
@@ -33,6 +33,7 @@ import org.bytesoft.common.utils.CommonUtils;
 import org.bytesoft.compensable.CompensableBeanFactory;
 import org.bytesoft.compensable.CompensableManager;
 import org.bytesoft.compensable.CompensableTransaction;
+import org.bytesoft.compensable.TransactionBeanFactory;
 import org.bytesoft.compensable.aware.CompensableBeanFactoryAware;
 import org.bytesoft.compensable.aware.CompensableEndpointAware;
 import org.bytesoft.compensable.logging.CompensableLogger;
@@ -54,6 +55,10 @@ public class CompensableCoordinator implements RemoteCoordinator, CompensableBea
 
 	@javax.inject.Inject
 	private CompensableBeanFactory beanFactory;
+
+	@javax.inject.Inject
+	private TransactionBeanFactory transactionBeanFactory;
+
 	private String endpoint;
 
 	private transient boolean ready = false;
@@ -331,7 +336,7 @@ public class CompensableCoordinator implements RemoteCoordinator, CompensableBea
 	public Xid[] recover(int flag) throws XAException {
 		this.checkParticipantReadyIfNecessary();
 
-		TransactionRepository repository = beanFactory.getTransactionRepository();
+		TransactionRepository repository = transactionBeanFactory.getTransactionRepository();
 		List<Transaction> transactionList = repository.getActiveTransactionList();
 
 		TransactionXid[] xidArray = new TransactionXid[transactionList.size()];
