+++ bytetcc-supports-springboot/src/main/java/org/bytesoft/bytetcc/supports/springboot/web/CompensableHandlerInterceptor.java
@@ -27,11 +27,7 @@ import org.bytesoft.bytejta.supports.rpc.TransactionResponseImpl;
 import org.bytesoft.bytetcc.supports.springboot.SpringBootBeanRegistry;
 import org.bytesoft.bytetcc.supports.springboot.controller.CompensableCoordinatorController;
 import org.bytesoft.common.utils.SerializeUtils;
-import org.bytesoft.compensable.Compensable;
-import org.bytesoft.compensable.CompensableBeanFactory;
-import org.bytesoft.compensable.CompensableManager;
-import org.bytesoft.compensable.CompensableTransaction;
-import org.bytesoft.compensable.TransactionContext;
+import org.bytesoft.compensable.*;
 import org.bytesoft.compensable.aware.CompensableEndpointAware;
 import org.bytesoft.transaction.supports.rpc.TransactionInterceptor;
 import org.slf4j.Logger;
@@ -90,7 +86,8 @@ public class CompensableHandlerInterceptor implements HandlerInterceptor, Compen
 
 		SpringBootBeanRegistry beanRegistry = SpringBootBeanRegistry.getInstance();
 		CompensableBeanFactory beanFactory = beanRegistry.getBeanFactory();
-		TransactionInterceptor transactionInterceptor = beanFactory.getTransactionInterceptor();
+		TransactionBeanFactory transactionBeanFactory = (TransactionBeanFactory) beanRegistry.getBeanFactory();
+		TransactionInterceptor transactionInterceptor = transactionBeanFactory.getTransactionInterceptor();
 
 		byte[] byteArray = transactionText == null ? new byte[0] : Base64.getDecoder().decode(transactionText);
 
@@ -156,7 +153,8 @@ public class CompensableHandlerInterceptor implements HandlerInterceptor, Compen
 		SpringBootBeanRegistry beanRegistry = SpringBootBeanRegistry.getInstance();
 		CompensableBeanFactory beanFactory = beanRegistry.getBeanFactory();
 		CompensableManager compensableManager = beanFactory.getCompensableManager();
-		TransactionInterceptor transactionInterceptor = beanFactory.getTransactionInterceptor();
+		TransactionBeanFactory transactionBeanFactory = (TransactionBeanFactory) beanRegistry.getBeanFactory();
+		TransactionInterceptor transactionInterceptor = transactionBeanFactory.getTransactionInterceptor();
 
 		CompensableTransaction compensable = compensableManager.getCompensableTransactionQuietly();
 		TransactionContext transactionContext = compensable.getTransactionContext();
